+++ History.md
@@ -3,6 +3,7 @@
 
   * `path` argument is required; there is no default icon.
   * Non-GET and HEAD requests are denied.
+  * Support OPTIONS method
 
 1.0.2 / 2014-03-16
 ==================